Reorganize imports
Mainly, move queryString.js imports out of the index.js.
