+++ indico/modules/rb/client/js/modules/blockings/index.js
@@ -10,8 +10,6 @@ import * as selectors from './selectors';
 
 export {default} from './BlockingList';
 export {default as BlockingModal} from './BlockingModal';
-export {default as BlockingPreloader} from './BlockingPreloader';
 export {default as reducer} from './reducers';
 export {default as modalHandlers} from './modals';
-export {queryStringReducer, routeConfig} from './queryString';
 export {actions, selectors};
