+++ indico/modules/rb/client/js/modules/bookRoom/actions.js
@@ -18,8 +18,7 @@ import {validateFilters} from '../../common/filters';
 import {preProcessParameters} from '../../util';
 import {actions as modalActions} from '../../modals';
 
-import {roomSearchActionsFactory} from '../../common/roomSearch/actions';
-import {ajax as ajaxFilterRules} from '../../common/roomSearch/serializers';
+import {roomSearchActionsFactory, ajaxRules as ajaxFilterRules} from '../../common/roomSearch';
 import {selectors as userSelectors} from '../../common/user';
 
 // Booking creation
