+++ indico/modules/rb/client/js/modules/bookRoom/reducers.js
@@ -11,7 +11,7 @@ import {camelizeKeys} from 'indico/utils/case';
 import {requestReducer} from 'indico/utils/redux';
 import * as actions from './actions';
 import * as globalActions from '../../actions';
-import {roomSearchReducerFactory} from '../../common/roomSearch';
+import {roomSearchReducerFactory} from '../../common/roomSearch/reducers';
 import {initialDatePickerState} from '../../common/timeline/reducers';
 
 export const initialTimelineState = {
