+++ indico/modules/rb/client/js/modules/calendar/index.js
@@ -9,5 +9,4 @@ import * as actions from './actions';
 
 export {default} from './Calendar';
 export {default as reducer} from './reducers';
-export {queryStringReducer, routeConfig} from './queryString';
 export {actions};
