+++ indico/modules/rb/client/js/modules/roomList/index.js
@@ -10,5 +10,4 @@ import * as selectors from './selectors';
 export {default} from './RoomList';
 export {default as roomFilterBarFactory} from './RoomFilterBar';
 export {default as reducer} from './reducers';
-export {routeConfig} from './queryString';
 export {selectors};
